+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 Starting from Rust port of the project, all changes will be put into this file.
 
 ## Unreleased (git master)
-
-Nothing yet!
+### Build
+- Pin `mime` and `windows-sys` dependencies
 
 ## [0.3.0] 2024-02-14
 ### New Features